Trac #32867: new spkg-configure.m4 for maxima.
This new spkg-configure.m4 checks for both a "maxima" executable, and
the usability of a "maxima" package by ECL. (The latter requires a
patched maxima until a new release with commit a0d7a43e523 is made.)
Notably absent for the moment is a version check, but that should be
okay: at the moment, the only system copies of maxima that support
being used as an ECL library are those that are patched, and the only
reason to patch them is if you intend to use them with SageMath. So in
short, if a distro patches their maxima to use it with sage, we expect
it to work with sage.

This commit also moves the SAGE_MAXIMA_FAS variable handling from
ECL's spkg-configure.m4 to maxima's (where it would have belonged in
the first place, if maxima had an spkg-configure.m4 at the time.)