src/sage/env.py: pass -A to the gap command
From the GAP documentation:

  By default, some needed and suggested GAP packages are loaded, if
  present, into the GAP session when it starts. This option disables
  (actually toggles) the loading of suggested packages, which can be
  useful for debugging or testing. The needed packages (and their
  needed packages, and so on) are loaded in any case.

With the possible exception of PolyCyclic, Sage doesn't need any of
those suggested packages, but GAP complains loudly if they're
missing. Passing "-A" to the GAP command eliminates those warnings and
allows us to pare down our GAP installation, or to use a more minimal
GAP from the system.