src/sage/libs/gap/sage.gaprc: load PolyCyclic if possible
Now that we're initalizing libgap with "-A", the PolyCyclic package
won't be loaded even if it is installed. We do want to use it if it's
available, though, so we call LoadPackage explicitly in sage.gaprc to
load it. This spews some informational messages if PolyCyclic is NOT
installed, but we can hide them by temporarily changing the
InfoWarning level.