src/sage/features/__init__.py: update PolyCyclic unhide() example
There's a test for unhide() in this module that assumes the PolyCyclic
GAP package will always be installed. That may not be the case when
now that we support using the system GAP. We update the text of the
example and add two "# optional" tags, one of which is needed to
hide the feature in the first place.