src/sage/graphs/generators/families.py: remove unpredictable doctest
One of our examples for MathonPseudocyclicStronglyRegularGraph()
constructs a matrix from a normal subgroup of the graph's automorphism
group. The normal_subgroups() method goes through GAP, and GAP's
NormalSubgroups() function has some freedom to choose representatives.
This leads to the entries of the matrix being unpredictable -- it
still satisfies the desired properties, however.

To avoid doctest failues due to this unpredictability, we no longer
print the entries of the matrix. The rest of the example is fine.